feat: add first swiftdata database schema
