+++ Wallet/Core/Models/SavedCredentialModel.swift
@@ -4,26 +4,7 @@
 
 import Foundation
 
-enum CredentialType: String, Codable, Sendable {
-  case pid = "urn:eudi:pid:1"
-}
-
-struct IssuerDisplay: Codable, Hashable, Sendable {
-  let name: String
-  let info: String?
-  let imageUrl: URL?
-}
-
-struct SavedCredential: Codable, Hashable, Sendable {
-  let issuer: IssuerDisplay
-  let compactSerialized: String
-  let claimDisplayNames: [String: String]
-  let claimsCount: Int
-  var issuedAt: Date = .now
-  let type: String
-  let displayData: CredentialDisplayData?
-}
-
-struct CredentialDisplayData: Codable, Hashable, Sendable {
-  let name: String?
-}
+typealias CredentialType = SchemaV1.CredentialType
+typealias IssuerDisplay = SchemaV1.IssuerDisplay
+typealias SavedCredential = SchemaV1.SavedCredential
+typealias CredentialDisplayData = SchemaV1.CredentialDisplayData
